# apps/api/src/identity.ts
import type { FastifyInstance, FastifyRequest } from "fastify";
import { ObjectId } from "mongodb";
import { z } from "zod";
import type { AppConfig } from "./config.js";
import type { AgentDocument, ApprovalDocument, Collections, PolicyDocument } from "./db.js";
import {
  authenticateAgentRequest,
  issueIdentityToken,
  type IdentityTokenMode
} from "./agent-auth.js";
import { AUDIT_ACTIONS, listAuditEntries, recordAudit, serializeAuditEntry } from "./audit.js";
import { ApiError } from "./errors.js";
import { slugify } from "./lib/slug.js";
import { normalizeEmail } from "./security.js";
import { getProvisioner } from "./provisioning.js";
import { registerEmailProvisioner } from "./email-provisioning.js";
import { defaultEmailPolicy } from "./policies.js";
import {
  getAgentPhoneCall,
  listAgentPhoneCalls,
  placeOutboundCall,
  serializePhoneCall
} from "./phone-service.js";
import {
  findLatestSmsCode,
  listAgentSmsConversation,
  sendAgentSmsWithPolicy,
  serializeSmsMessage
} from "./sms-service.js";

type ToolName = "email" | "phone";
type PermissionName = "email.send" | "phone.call";

export interface AgentIdentity {
  id: string;
  name: string;
  status: "active" | "revoked";
  permissions: Record<PermissionName, boolean> & {
    requiresHumanApproval: boolean;
  };
}

function agentIdentityView(agent: AgentDocument): AgentIdentity {
  return {
    id: agent._id.toHexString(),
    name: agent.name,
    status: agent.status === "active" ? "active" : "revoked",
    permissions: {
      "email.send": agent.capabilities.email,
      "phone.call": agent.capabilities.phone,
      requiresHumanApproval: agent.approvalMode !== "autonomous"
    }
  };
}

async function loadAgentIdentityFromRequest(
  request: FastifyRequest,
  collections: Collections
): Promise<AgentIdentity | null> {
  const agentContext = await authenticateAgentRequest(request, collections);
  return agentContext ? agentIdentityView(agentContext.agent) : null;
}

const initIdentitySchema = z.object({
  agent_name: z.string().min(1).max(80),
  agent_runtime: z.string().min(1).max(80).default("openclaw"),
  use_case: z.string().min(1).max(120).default("automation"),
  owner_email: z.string().email().optional(),
  tools: z
    .array(z.enum(["email", "phone"]))
    .min(1)
    .default(["email", "phone"]),
  permissions: z
    .object({
      "email.send": z.boolean().optional(),
      "phone.call": z.boolean().optional(),
      requires_human_approval: z.boolean().optional()
    })
    .optional()
});

const phoneCallSchema = z.object({
  to: z.string().min(3).max(40),
  script: z.string().min(1).max(5000).optional(),
  task: z.string().min(1).max(5000).optional(),
  context: z.string().max(8000).optional(),
  recipientName: z.string().max(200).optional(),
  approved: z.boolean().optional()
}).refine((payload) => payload.script || payload.task, { message: "script or task is required" });

const phoneCallsQuerySchema = z.object({
  cursor: z.string().optional()
});

const approvalQuerySchema = z.object({
  wait: z.coerce.number().int().min(1).max(300).optional(),
  mode: z.enum(["async"]).optional()
});

const smsSendSchema = z.object({
  to: z.string().min(3).max(40),
  body: z.string().min(1).max(1600),
  idempotencyKey: z.string().min(1).max(120).optional()
});

const smsListQuerySchema = z.object({
  with: z.string().min(3).max(40),
  cursor: z.string().optional()
});

const smsLatestCodeQuerySchema = z.object({
  from: z.string().min(3).max(40).optional(),
  since: z.string().datetime().optional()
});

const auditRecentQuerySchema = z.object({
  limit: z.coerce.number().int().min(1).max(100).default(25)
});

export function registerIdentityRoutes(app: FastifyInstance, collections: Collections, config: AppConfig) {
  registerEmailProvisioner(collections, config);

  app.get("/api/v1/agent/whoami", async (request) => {
    const agentContext = await authenticateAgentRequest(request, collections);
    if (!agentContext) {
      throw new ApiError(401, "unauthorized", "missing or invalid identity token");
    }
    const [emailAccount, phoneNumber] = await Promise.all([
      collections.emailAccounts.findOne({ agentId: agentContext.agent._id, status: "active" }),
      collections.phoneNumbers.findOne({ agentId: agentContext.agent._id, status: "active" })
    ]);
    return {
      agent_id: agentContext.agent._id.toHexString(),
      name: agentContext.agent.name,
      slug: agentContext.agent.slug,
      status: agentContext.agent.status,
      runtime: agentContext.agent.runtime ?? null,
      capabilities: agentContext.agent.capabilities,
      approval_mode: agentContext.agent.approvalMode,
      email: emailAccount?.address ?? null,
      phone: phoneNumber?.e164 ?? null,
      token_prefix: agentContext.token.prefix
    };
  });

  app.get("/api/v1/agent/audit/recent", async (request) => {
    const agentContext = await authenticateAgentRequest(request, collections);
    if (!agentContext) {
      throw new ApiError(401, "unauthorized", "missing or invalid identity token");
    }
    const query = auditRecentQuerySchema.parse(request.query ?? {});
    const page = await listAuditEntries(collections, { agentId: agentContext.agent._id, limit: query.limit });
    return {
      entries: page.entries.map(serializeAuditEntry),
      next_cursor: page.nextCursor
    };
  });

  app.post("/api/identity/init", async (request, reply) => {
    const payload = initIdentitySchema.parse(request.body ?? {});
    const tools = [...new Set(payload.tools)] as ToolName[];

    let ownerUserId: ObjectId | null = null;
    if (payload.owner_email) {
      const owner = await collections.users.findOne({ email: normalizeEmail(payload.owner_email) });
      if (!owner) {
        throw new ApiError(404, "not_found", "owner_email does not match an existing user");
      }
      ownerUserId = owner._id;
    }

    const now = new Date();
    const agent: AgentDocument = {
      _id: new ObjectId(),
      ownerUserId,
      name: payload.agent_name.trim(),
      slug: await reserveAgentSlug(collections, ownerUserId, payload.agent_name),
      status: "active",
      description: payload.use_case.trim(),
      runtime: normalizeRuntime(payload.agent_runtime),
      capabilities: { email: tools.includes("email"), phone: tools.includes("phone") },
      approvalMode: (payload.permissions?.requires_human_approval ?? true) ? "always" : "autonomous",
      createdAt: now,
      updatedAt: now
    };
    await collections.agents.insertOne(agent);

    const policy: PolicyDocument = {
      _id: new ObjectId(),
      agentId: agent._id,
      email: defaultEmailPolicy(agent.approvalMode),
      phone: {},
      createdAt: now,
      updatedAt: now
    };
    await collections.policies.insertOne(policy);
    let emailAddress: string | null = null;
    if (agent.capabilities.email) {
      try {
        await getProvisioner("email").provision(agent);
        emailAddress = (await collections.emailAccounts.findOne({ agentId: agent._id, status: "active" }))?.address ?? null;
      } catch (error) {
        await Promise.all([
          collections.agents.deleteOne({ _id: agent._id }),
          collections.policies.deleteMany({ agentId: agent._id }),
          collections.emailAccounts.deleteMany({ agentId: agent._id })
        ]);
        throw error;
      }
    }

    const { plaintext } = await issueIdentityToken(collections, agent._id, "default", {
      mode: identityTokenMode(config)
    });

    await recordAudit(collections, {
      agentId: agent._id,
      ownerUserId,
      actor: "system",
      action: AUDIT_ACTIONS.identity.init,
      status: "allowed",
      detail: `${agent.name} initialized for ${agent.runtime}.`
    });

    const identity = agentIdentityView(agent);
    return reply.code(201).send({
      agent_id: identity.id,
      identity_token: plaintext,
      status: agent.status,
      runtime: agent.runtime,
      use_case: agent.description,
      // Card is deferred.
      email: emailAddress,
      phone: null,
      payment: null,
      tools,
      permissions: serializePermissions(identity),
      openclaw_env: {
        BARKAN_API_URL: config.PUBLIC_API_URL,
        BARKAN_IDENTITY_TOKEN: plaintext
      },
      tool_endpoints: {
        email_send: `${config.PUBLIC_API_URL}/api/v1/agent/email/send`,
        email_threads: `${config.PUBLIC_API_URL}/api/v1/agent/email/threads`,
        email_activity: `${config.PUBLIC_API_URL}/api/identity/${identity.id}/email-activity`,
        phone_call: `${config.PUBLIC_API_URL}/api/v1/agent/phone/call`,
        phone_sms: `${config.PUBLIC_API_URL}/api/v1/agent/phone/sms`,
        audit_log: `${config.PUBLIC_API_URL}/api/identity/${identity.id}/audit-log`,
        token_rotate: `${config.PUBLIC_API_URL}/api/identity/tokens/rotate`
      }
    });
  });

  app.post("/api/v1/agent/phone/call", async (request) => {
    const agentContext = await authenticateAgentRequest(request, collections);
    if (!agentContext) {
      throw new ApiError(401, "unauthorized", "missing or invalid identity token");
    }
    const payload = phoneCallSchema.parse(request.body ?? {});
    const query = approvalQuerySchema.parse(request.query ?? {});

    const result = await placeOutboundCall(collections, config, {
      agent: agentContext.agent,
      toNumber: payload.to,
      task: payload.task ?? payload.script ?? "",
      context: payload.context,
      recipientName: payload.recipientName
    }, { async: query.mode === "async", waitMs: query.wait ? query.wait * 1000 : undefined });
    if ("approvalRequired" in result) {
      return serializeApprovalPending(result.approval, result.decision);
    }
    return {
      call_id: result.callId,
      status: result.status,
      from: result.from,
      to: result.to
    };
  });

  app.get("/api/v1/agent/phone/calls", async (request) => {
    const agentContext = await authenticateAgentRequest(request, collections);
    if (!agentContext) {
      throw new ApiError(401, "unauthorized", "missing or invalid identity token");
    }
    const query = phoneCallsQuerySchema.parse(request.query ?? {});
    const page = await listAgentPhoneCalls(collections, agentContext.agent, query.cursor);
    return {
      calls: page.calls.map(serializePhoneCall),
      next_cursor: page.nextCursor
    };
  });

  app.get("/api/v1/agent/phone/number", async (request) => {
    const agentContext = await authenticateAgentRequest(request, collections);
    if (!agentContext) {
      throw new ApiError(401, "unauthorized", "missing or invalid identity token");
    }
    const phoneNumber = await collections.phoneNumbers.findOne({ agentId: agentContext.agent._id, status: "active" });
    if (!phoneNumber) {
      throw new ApiError(409, "policy_blocked", "phone capability not provisioned");
    }
    return {
      e164: phoneNumber.e164,
      country: phoneNumber.country,
      capabilities: { voice: phoneNumber.capabilitiesVoice, sms: phoneNumber.capabilitiesSms },
      status: phoneNumber.status
    };
  });

  app.get("/api/v1/agent/phone/calls/:callId", async (request) => {
    const agentContext = await authenticateAgentRequest(request, collections);
    if (!agentContext) {
      throw new ApiError(401, "unauthorized", "missing or invalid identity token");
    }
    const { callId } = request.params as { callId: string };
    return { call: serializePhoneCall(await getAgentPhoneCall(collections, agentContext.agent, callId)) };
  });

  app.post("/api/v1/agent/phone/sms", async (request) => {
    const agentContext = await authenticateAgentRequest(request, collections);
    if (!agentContext) {
      throw new ApiError(401, "unauthorized", "missing or invalid identity token");
    }
    const payload = smsSendSchema.parse(request.body ?? {});
    const query = approvalQuerySchema.parse(request.query ?? {});
    const result = await sendAgentSmsWithPolicy(collections, config, {
      agent: agentContext.agent,
      to: payload.to,
      body: payload.body,
      idempotencyKey: payload.idempotencyKey
    }, { async: query.mode === "async", waitMs: query.wait ? query.wait * 1000 : undefined });
    if ("approvalRequired" in result) {
      return serializeApprovalPending(result.approval, result.decision);
    }
    return { message: serializeSmsMessage(result) };
  });

  app.get("/api/v1/agent/phone/sms", async (request) => {
    const agentContext = await authenticateAgentRequest(request, collections);
    if (!agentContext) {
      throw new ApiError(401, "unauthorized", "missing or invalid identity token");
    }
    const query = smsListQuerySchema.parse(request.query ?? {});
    const page = await listAgentSmsConversation(collections, agentContext.agent, { with: query.with, cursor: query.cursor });
    return {
      messages: page.messages.map(serializeSmsMessage),
      next_cursor: page.nextCursor
    };
  });

  app.get("/api/v1/agent/phone/sms/latest-code", async (request) => {
    const agentContext = await authenticateAgentRequest(request, collections);
    if (!agentContext) {
      throw new ApiError(401, "unauthorized", "missing or invalid identity token");
    }
    const query = smsLatestCodeQuerySchema.parse(request.query ?? {});
    const result = await findLatestSmsCode(collections, agentContext.agent, {
      from: query.from,
      since: query.since ? new Date(query.since) : null
    });
    return { code: result.code, receivedAt: result.receivedAt.toISOString(), from: result.from };
  });

  app.get("/api/identity/:agentId/audit-log", async (request, reply) => {
    const agentContext = await authenticateAgentRequest(request, collections);
    if (!agentContext) {
      throw new ApiError(401, "unauthorized", "missing or invalid identity token");
    }

    const { agentId } = request.params as { agentId: string };
    if (agentContext.agent._id.toHexString() !== agentId) {
      throw new ApiError(403, "forbidden", "identity token does not match requested agent");
    }

    const { entries } = await listAuditEntries(collections, {
      agentId: agentContext.agent._id,
      limit: 100
    });

    return {
      agent_id: agentId,
      status: agentContext.agent.status,
      audit_log: entries.map((entry) => {
        const serialized = serializeAuditEntry(entry);
        return {
          id: serialized.id,
          agent_id: serialized.agent_id,
          action: serialized.action,
          status: serialized.status,
          detail: serialized.detail,
          created_at: serialized.created_at
        };
      })
    };
  });

  app.post("/api/identity/revoke", async (request, reply) => {
    const agentContext = await authenticateAgentRequest(request, collections);
    if (!agentContext) {
      throw new ApiError(401, "unauthorized", "missing or invalid identity token");
    }

    await collections.identityTokens.updateOne(
      { _id: agentContext.token._id },
      { $set: { status: "revoked", updatedAt: new Date() } }
    );
    await recordAudit(collections, {
      agentId: agentContext.agent._id,
      ownerUserId: agentContext.agent.ownerUserId,
      actor: "agent",
      action: AUDIT_ACTIONS.identity.revoke,
      status: "allowed",
      detail: `Identity token ${agentContext.token.prefix}… revoked.`
    });
    return {
      ok: true,
      agent_id: agentContext.agent._id.toHexString(),
      status: "revoked"
    };
  });

  app.post("/api/identity/tokens/rotate", async (request, reply) => {
    const agentContext = await authenticateAgentRequest(request, collections);
    if (!agentContext) {
      throw new ApiError(401, "unauthorized", "missing or invalid identity token");
    }

    // Issue the replacement before revoking the old token so the agent is
    // never left without a working credential.
    const { plaintext, tokenDoc } = await issueIdentityToken(
      collections,
      agentContext.agent._id,
      agentContext.token.name,
      { mode: identityTokenMode(config) }
    );
    await collections.identityTokens.updateOne(
      { _id: agentContext.token._id },
      { $set: { status: "revoked", updatedAt: new Date() } }
    );
    await recordAudit(collections, {
      agentId: agentContext.agent._id,
      ownerUserId: agentContext.agent.ownerUserId,
      actor: "agent",
      action: AUDIT_ACTIONS.identity.tokenRotate,
      status: "allowed",
      detail: `Identity token ${agentContext.token.prefix}… rotated to ${tokenDoc.prefix}….`
    });
    return {
      ok: true,
      agent_id: agentContext.agent._id.toHexString(),
      identity_token: plaintext,
      token_prefix: tokenDoc.prefix
    };
  });
}

function serializeApprovalPending(approval: ApprovalDocument, decision: "pending" | "timeout" | "expired") {
  return {
    ok: false,
    status: "approval_required",
    decision,
    approval_id: approval._id.toHexString(),
    approval: {
      id: approval._id.toHexString(),
      status: approval.status,
      payloadSummary: approval.payloadSummary,
      executionResult: approval.executionResult ?? null,
      executionError: approval.executionError ?? null
    }
  };
}

export async function reserveAgentSlug(
  collections: Collections,
  ownerUserId: ObjectId | null,
  agentName: string
): Promise<string> {
  const baseSlug = slugify(agentName);
  let candidate = baseSlug;
  let suffix = 2;
  while (await collections.agents.findOne({ ownerUserId, slug: candidate })) {
    candidate = `${baseSlug}-${suffix}`;
    suffix += 1;
  }
  return candidate;
}

function normalizeRuntime(runtime: string): NonNullable<AgentDocument["runtime"]> {
  const normalized = runtime.trim().toLowerCase();
  if (normalized === "openclaw" || normalized === "hermes" || normalized === "api") {
    return normalized;
  }
  return "other";
}

export function identityTokenMode(config: AppConfig): IdentityTokenMode {
  return config.PROVIDER_MODE_EMAIL === "mock" && config.PROVIDER_MODE_PHONE === "mock" ? "test" : "live";
}

function serializePermissions(identity: AgentIdentity) {
  return {
    "email.send": identity.permissions["email.send"],
    "phone.call": identity.permissions["phone.call"],
    requires_human_approval: identity.permissions.requiresHumanApproval
  };
}